fix circular dependencies, add madge to github actions
