+++ .github/workflows/checks.yml
@@ -20,5 +20,7 @@ jobs:
         run: yarn install --frozen-lockfile
       - name: Run ESLint
         run: yarn atlas:lint
+      - name: Run madge
+        run: yarn atlas:madge:circular
       - name: Run Tests
         run: yarn atlas:test
