+++ packages/atlas/package.json
@@ -24,7 +24,8 @@
     "svgs:icons": "yarn svgs:icons-import && yarn svgs:icons-generate",
     "graphql": "rimraf src/api/queries/__generated__ && graphql-codegen --config codegen.config.yml -r dotenv/config",
     "test": "vitest run",
-    "test:dev": "vitest dev"
+    "test:dev": "vitest dev",
+    "madge:circular": "madge --circular ./src/* --ts-config ./tsconfig.json --extensions ts,tsx"
   },
   "dependencies": {
     "@apollo/client": "^3.5.8",
@@ -56,6 +57,7 @@
     "immer": "^9.0.12",
     "localforage": "^1.10.0",
     "lodash-es": "^4.17.21",
+    "madge": "^5.0.1",
     "multihashes": "^4.0.3",
     "postcss-syntax": "^0.36.2",
     "rc-slider": "^9.7.5",
