+++ packages/atlas/src/components/EmptyFallback/EmptyFallback.styles.ts
@@ -3,7 +3,7 @@ import styled from '@emotion/styled'
 import { Text } from '@/components/Text'
 import { media, sizes } from '@/styles'
 
-import { EmptyFallbackSizes } from './EmptyFallback'
+export type EmptyFallbackSizes = 'small' | 'large'
 
 type ContainerProps = {
   variant?: EmptyFallbackSizes
