+++ packages/atlas/src/components/_buttons/CallToActionButton/CallToActionButton.tsx
@@ -5,9 +5,7 @@ import { SvgActionChevronR } from '@/components/_icons'
 import { useMediaMatch } from '@/hooks/useMediaMatch'
 import { getLinkPropsFromTo } from '@/utils/button'
 
-import { BodyWrapper, ContentWrapper, IconWrapper, StyledContainer } from './CallToActionButton.styles'
-
-export type ColorVariants = 'red' | 'green' | 'yellow' | 'blue' | 'lightBlue' | 'white'
+import { BodyWrapper, ColorVariants, ContentWrapper, IconWrapper, StyledContainer } from './CallToActionButton.styles'
 
 export type CallToActionButtonProps = {
   to?: To
