+++ packages/atlas/src/components/_nft/NftWidget/NftWidget.tsx
@@ -34,12 +34,10 @@ import {
   OwnerAvatar,
   OwnerHandle,
   OwnerLabel,
+  Size,
   TopBidderContainer,
   TopBidderTokenContainer,
-  sizeObj,
 } from './NftWidget.styles'
-
-export type Size = keyof typeof sizeObj
 
 export type Auction = {
   status: 'auction'
