+++ packages/atlas/src/views/viewer/ChannelView/ChannelSearch.tsx
@@ -3,8 +3,7 @@ import React, { FormEvent, useCallback, useEffect, useRef } from 'react'
 import { SvgActionSearch } from '@/components/_icons'
 
 import { SearchButton, SearchContainerForm, StyledTextField } from './ChannelSearch.styles'
-
-import { TABS } from '.'
+import { TABS } from './utils'
 
 type SearchProps = {
   isSearchInputOpen: boolean
