+++ packages/atlas/src/views/viewer/ChannelView/utils.ts
@@ -1,0 +1,1 @@
+export const TABS = ['Videos', 'NFTs', 'Information'] as const
